Move set_cpu_affinity to rt::setup module
