+++ src/rt/worker.rs
@@ -25,6 +25,7 @@ use mio::{Poll, Registry, Token};
 
 use crate::rt::local::waker::{self, WakerId};
 use crate::rt::local::{Control, Runtime, WAKER};
+use crate::rt::setup::set_cpu_affinity;
 use crate::rt::thread_waker::ThreadWaker;
 use crate::rt::{self, shared, ProcessId, RuntimeRef, Signal};
 use crate::trace;
@@ -186,51 +187,4 @@ fn main(
 
     // All setup is done, so we're ready to run the event loop.
     runtime.run_event_loop().map_err(rt::Error::worker)
-}
-
-/// Set thread's CPU affinity.
-fn set_cpu_affinity(worker_id: NonZeroUsize) -> Option<usize> {
-    #[cfg(not(target_os = "linux"))]
-    {
-        let _ = worker_id; // Silence unused variables warnings.
-        None
-    }
-
-    #[cfg(target_os = "linux")]
-    {
-        let cpu = worker_id.get() - 1; // Worker ids start at 1, cpus at 0.
-        let cpu_set = cpu_set(cpu);
-        match set_affinity(&cpu_set) {
-            Ok(()) => {
-                log::debug!("worker thread using CPU '{}'", cpu);
-                Some(cpu)
-            }
-            Err(err) => {
-                log::warn!("error setting CPU affinity: {}", err);
-                None
-            }
-        }
-    }
-}
-
-/// Create a cpu set that may only run on `cpu`.
-#[cfg(target_os = "linux")]
-fn cpu_set(cpu: usize) -> libc::cpu_set_t {
-    let mut cpu_set = unsafe { std::mem::zeroed() };
-    unsafe { libc::CPU_ZERO(&mut cpu_set) };
-    unsafe { libc::CPU_SET(cpu % libc::CPU_SETSIZE as usize, &mut cpu_set) };
-    cpu_set
-}
-
-/// Set the affinity of this thread to the `cpu_set`.
-#[cfg(target_os = "linux")]
-fn set_affinity(cpu_set: &libc::cpu_set_t) -> io::Result<()> {
-    let thread = unsafe { libc::pthread_self() };
-    let res =
-        unsafe { libc::pthread_setaffinity_np(thread, std::mem::size_of_val(cpu_set), cpu_set) };
-    if res == 0 {
-        Ok(())
-    } else {
-        Err(io::Error::last_os_error())
-    }
 }
